fix(PERF-A01): cast observation start timestamp

+++ starter/src/main/java/cn/jia/config/ApiPerformanceMetricsConfig.java
@@ -133,7 +133,7 @@ public class ApiPerformanceMetricsConfig {
 
         @Override
         public void onStop(ServerRequestObservationContext context) {
-            Long startedAt = context.remove(START_NANOS_KEY);
+            Long startedAt = (Long) context.remove(START_NANOS_KEY);
             if (startedAt == null) {
                 return;
             }
